refactor(utility): move serde functions
