+++ rust/src/utility/mod.rs
@@ -1,3 +1,4 @@
 pub mod compression;
 pub mod functions;
+pub mod serde;
 pub mod store;
